refactor (match): add match to it's own app
