+++ backend/django_backend/match/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class MatchConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'match'
