+++ backend/django_backend/match/views.py
@@ -1,0 +1,34 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import status
+from .models import Match
+from .serializers import MatchSerializer
+from django.contrib.auth.models import User
+from user.permissions import IsAuthenticatedOrCreateOnly
+
+class MatchListView(APIView):
+    permission_classes = [IsAuthenticatedOrCreateOnly]
+
+    def get(self, request, format=None):
+        matches = Match.objects.all()
+        serializer = MatchSerializer(matches, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, format=None):
+        serializer = MatchSerializer(data=request.data)
+        if(serializer.is_valid()):
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class PlayerMatchListView(APIView):
+    permissions_classes = [IsAuthenticatedOrCreateOnly]
+
+    def get(self, request, user_id, formant=None):
+        try:
+            player = User.objects.get(id=user_id)
+        except User.DoesNotExist:
+            return Response({"error": "User not found."}, status=status.HTTP_404_NOT_FOUND)
+        matches = Match.objects.filter(player1=player) | Match.objects.filter(player2=player)
+        serializer = MatchSerializer(matches, many=True)
+        return Response(serializer.data)
